requirements.txt
progressbar2
argparse
scikit-learn
torchvision
torchsummary
matplotlib
scipy
av